fix(tests): rename Functionnal → Functional (typo)
Fix the typo in directory name and update all namespaces and
phpunit.xml.dist accordingly.
